move removeUserFromOrg into a single test file using new testing method
